extract shared BinaryHeap for the search planners
KlondikePlanner and FreeCellSolver carried byte-identical private binary
heap implementations, and the upcoming Yukon planner would have added a
third. Behavior-preserving: identical comparator and algorithm, verified
by identical solve/win-rate probe results pre- and post-migration.
